Crate narwhal_types

Modules

error

Macros

bail

ensure

Structs

BatchDigest

BatchMessage

BatchV1

BatchV2

CertificateDigest

CertificateV1

CertificateV2

CommittedSubDag

CommittedSubDagShell

ConditionalBroadcastReceiver

ConsensusCommitV2

ConsensusOutput

The output of Consensus, which includes all the batches for each certificate in the sub dag It is sent to the ExecutionState handle_consensus_transactions

ConsensusOutputDigest

Empty

Empty message for when we don’t have anything to return

FetchBatchesRequest

Used by the primary to request that the worker fetch the missing batches and reply with all of the content.

FetchBatchesResponse

All batches requested by the primary.

FetchCertificatesRequest

Used by the primary to fetch certificates from other primaries.

FetchCertificatesResponse

Used by the primary to reply to FetchCertificatesRequest.

HeaderDigest

HeaderV1

HeaderV1Builder

Builder for HeaderV1.

Metadata

MetadataV1

MockPrimaryToPrimary

Generated trait containing RPC methods that should be implemented for use with PrimaryToPrimaryServer.

MockPrimaryToWorker

Generated trait containing RPC methods that should be implemented for use with PrimaryToWorkerServer.

MockWorkerToPrimary

Generated trait containing RPC methods that should be implemented for use with WorkerToPrimaryServer.

MockWorkerToWorker

Generated trait containing RPC methods that should be implemented for use with WorkerToWorkerServer.

PreSubscribedBroadcastSender

PreSubscribedBroadcastSender is a wrapped Broadcast channel that limits subscription to initialization time. This is designed to be used for cancellation signal to all the components, and the limitation is intended to prevent a component missing the shutdown signal due to a subscription that happens after the shutdown signal was sent. The receivers have a special peek method which can be used to conditionally check for shutdown signal on the channel.

PrimaryToPrimaryClient

PrimaryToPrimaryServer

PrimaryToWorkerClient

PrimaryToWorkerServer

RandomnessRound

DEPRECATED. Can’t be deleted until tables that use this are removed.

ReputationScores

RequestBatchesRequest

Used by primary to bulk request batches from workers local store.

RequestBatchesResponse

RequestVoteRequest

Used by the primary to request a vote from other primaries on newly produced headers.

RequestVoteResponse

Used by the primary to reply to RequestVoteRequest.

SendCertificateRequest

Request for broadcasting certificates to peers.

SendCertificateResponse

Response from peers after receiving a certificate.

TransactionProto

TransactionsClient

TransactionsServer

VoteDigest

VoteInfoV1

VoteV1

WorkerBatchMessage

Used by workers to send a new batch.

WorkerOthersBatchMessage

Used by worker to inform primary it received a batch from another authority.

WorkerOwnBatchMessage

Used by worker to inform primary it sealed a new batch.

WorkerSynchronizeMessage

Used by the primary to request that the worker sync the target missing batches.

WorkerToPrimaryClient

WorkerToPrimaryServer

WorkerToWorkerClient

WorkerToWorkerServer

Enums

Batch

Certificate

ConsensusCommit

DigestError

Header

HeaderV1BuilderError

Error type for HeaderV1Builder

SignatureVerificationState

VersionedMetadata

Vote

A Vote on a Header is a claim by the voting authority that all payloads and the full history of Certificates included in the Header are available.

VoteInfo

Traits

BatchAPI

CertificateAPI

HeaderAPI

MetadataAPI

PrimaryToPrimary

Generated trait containing RPC methods that should be implemented for use with PrimaryToPrimaryServer.

PrimaryToWorker

Generated trait containing RPC methods that should be implemented for use with PrimaryToWorkerServer.

Timestamp

Transactions

Generated trait containing gRPC methods that should be implemented for use with TransactionsServer.

VoteAPI

VoteInfoAPI

WorkerToPrimary

Generated trait containing RPC methods that should be implemented for use with WorkerToPrimaryServer.

WorkerToWorker

Generated trait containing RPC methods that should be implemented for use with WorkerToWorkerServer.

Functions

now

validate_batch_version

validate_received_certificate_version

Type Aliases

Round

The round number.

SequenceNumber

A global sequence number assigned to every CommittedSubDag.

ShutdownToken

Shutdown token dropped when a task is properly shut down.

TimestampMs

The epoch UNIX timestamp in milliseconds

Transaction

TxResponse